web:LB-10: откат userId к типу Int

+++ logbot-web/composeApp/src/wasmJsMain/kotlin/com/vk/logbot/web/model/UserInfo.kt
@@ -4,6 +4,6 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class UserInfo(
-    val telegramId: Long,
+    val telegramId: Int,
     val idCompany: String
 )
